Fix failing tests, format responding CSV
